Enable StrongName verification in SignCheck (#15602)
