+++ Arcade.sln
@@ -151,6 +151,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.MacOsPkg.C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.MacOsPkg.Tests", "src\Microsoft.DotNet.MacOsPkg.Tests\Microsoft.DotNet.MacOsPkg.Tests.csproj", "{1F5118A8-A5C5-4D18-AF34-FFB60FECCD45}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.StrongName", "src\Microsoft.DotNet.StrongName\Microsoft.DotNet.StrongName.csproj", "{B65809A4-C6DB-4994-BBDE-D5B8396958A6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1001,6 +1003,18 @@ Global
 		{1F5118A8-A5C5-4D18-AF34-FFB60FECCD45}.Release|x64.Build.0 = Release|Any CPU
 		{1F5118A8-A5C5-4D18-AF34-FFB60FECCD45}.Release|x86.ActiveCfg = Release|Any CPU
 		{1F5118A8-A5C5-4D18-AF34-FFB60FECCD45}.Release|x86.Build.0 = Release|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Debug|x64.Build.0 = Debug|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Debug|x86.Build.0 = Debug|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Release|x64.ActiveCfg = Release|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Release|x64.Build.0 = Release|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Release|x86.ActiveCfg = Release|Any CPU
+		{B65809A4-C6DB-4994-BBDE-D5B8396958A6}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
