+++ src/Microsoft.DotNet.SignTool/Microsoft.DotNet.SignTool.csproj
@@ -37,6 +37,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Microsoft.DotNet.Build.Tasks.Installers\Microsoft.DotNet.Build.Tasks.Installers.csproj" />
+    <ProjectReference Include="..\Microsoft.DotNet.StrongName\Microsoft.DotNet.StrongName.csproj" />
   </ItemGroup>
 
 </Project>
